+++ src/Cryptography/HmacSha256.cs
@@ -64,7 +64,7 @@ namespace NSec.Cryptography
         }
 
         internal override bool FinalizeAndTryVerifyCore(
-            ref IncrementalMac.State state,
+            ref IncrementalMacState state,
             ReadOnlySpan<byte> mac)
         {
             Debug.Assert(mac.Length <= crypto_auth_hmacsha256_BYTES);
@@ -79,7 +79,7 @@ namespace NSec.Cryptography
         }
 
         internal override void FinalizeCore(
-            ref IncrementalMac.State state,
+            ref IncrementalMacState state,
             Span<byte> mac)
         {
             Debug.Assert(mac.Length == crypto_auth_hmacsha256_BYTES);
@@ -95,7 +95,7 @@ namespace NSec.Cryptography
         internal override void InitializeCore(
             SecureMemoryHandle keyHandle,
             int macSize,
-            out IncrementalMac.State state)
+            out IncrementalMacState state)
         {
             Debug.Assert(keyHandle != null);
             Debug.Assert(macSize == crypto_auth_hmacsha256_BYTES);
@@ -140,7 +140,7 @@ namespace NSec.Cryptography
         }
 
         internal override void UpdateCore(
-            ref IncrementalMac.State state,
+            ref IncrementalMacState state,
             ReadOnlySpan<byte> data)
         {
             crypto_auth_hmacsha256_update(ref state.hmacsha256, in MemoryMarshal.GetReference(data), (ulong)data.Length);
