+++ src/Cryptography/Sha512.cs
@@ -44,7 +44,7 @@ namespace NSec.Cryptography
         }
 
         internal override bool FinalizeAndTryVerifyCore(
-            ref IncrementalHash.State state,
+            ref IncrementalHashState state,
             ReadOnlySpan<byte> hash)
         {
             Debug.Assert(hash.Length <= crypto_hash_sha512_BYTES);
@@ -59,7 +59,7 @@ namespace NSec.Cryptography
         }
 
         internal override void FinalizeCore(
-            ref IncrementalHash.State state,
+            ref IncrementalHashState state,
             Span<byte> hash)
         {
             Debug.Assert(hash.Length == crypto_hash_sha512_BYTES);
@@ -69,7 +69,7 @@ namespace NSec.Cryptography
 
         internal override void InitializeCore(
             int hashSize,
-            out IncrementalHash.State state)
+            out IncrementalHashState state)
         {
             Debug.Assert(hashSize == crypto_hash_sha512_BYTES);
 
@@ -77,7 +77,7 @@ namespace NSec.Cryptography
         }
 
         internal override void UpdateCore(
-            ref IncrementalHash.State state,
+            ref IncrementalHashState state,
             ReadOnlySpan<byte> data)
         {
             crypto_hash_sha512_update(ref state.sha512, in MemoryMarshal.GetReference(data), (ulong)data.Length);
